HDF5 version of zvals_trials
zvals_trial.mat is difficult to read without Matlab despite being in HDF5 format. Specifically the table in cluster_lookup is not readable with standard HDF5 readers. Attached is a version of the file in standard HDF5 format that's readable with any HDF5 reader or hdf5 library (such as h5py in python).